Move cache from lib/sqlbuilder to internal/.
