+++ internal/cache/hashstructure/LICENSE
