+++ internal/cache/hashstructure/hashstructure.go
