+++ internal/cache/hashstructure/hashstructure_test.go
