+++ internal/cache/interface.go
