+++ lib/sqlbuilder/exql/default.go
@@ -1,7 +1,7 @@
 package exql
 
 import (
-	"upper.io/db.v2/lib/sqlbuilder/cache"
+	"upper.io/db.v2/internal/cache"
 )
 
 const (
