+++ lib/sqlbuilder/template_test.go
@@ -1,7 +1,7 @@
 package builder
 
 import (
-	"upper.io/db.v2/lib/sqlbuilder/cache"
+	"upper.io/db.v2/internal/cache"
 	"upper.io/db.v2/lib/sqlbuilder/exql"
 )
