+++ mysql/template.go
@@ -22,7 +22,7 @@
 package mysql
 
 import (
-	"upper.io/db.v2/lib/sqlbuilder/cache"
+	"upper.io/db.v2/internal/cache"
 	"upper.io/db.v2/lib/sqlbuilder/exql"
 )
